requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
google-generativeai>=0.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
httpx>=0.25.0
fake-useragent>=1.4.0
python-docx>=1.1.0
pyyaml>=6.0.1
click>=8.1.0
rich>=13.7.0
tenacity>=8.2.0
